requirements.txt
absl-py==1.1.0
albumentations==1.3.0
astunparse==1.6.3
beautifulsoup4==4.11.1
blinker==1.6.2
cachetools==5.0.0
category-encoders==2.4.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==2.0.12
clang==5.0
click==8.1.3
cycler==0.11.0
dlib==19.24.0
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
einops==0.4.1
filelock==3.7.1
Flask==2.3.2
flatbuffers==1.12
fonttools==4.32.0
gast==0.4.0
gdown==4.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.5
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.44.0
h5py==3.1.0
idna==3.4
imageio==2.16.2
imbalanced-learn==0.9.0
imblearn==0.0
importlib-metadata==4.11.3
imutils==0.5.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kaggle==1.5.12
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
lightgbm==3.3.2
lxml==4.9.2
Markdown==3.3.6
MarkupSafe==2.1.2
matplotlib==3.5.2
mtcnn==0.1.1
mxnet==1.9.1
networkx==2.8
numpy==1.24.2
oauthlib==3.2.0
onnxruntime==1.11.1
opencv-python==4.5.5.64
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pathtools==0.1.2
patsy==0.5.2
Pillow==9.4.0
plyfile==0.7.4
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyg-lib==0.1.0+pt112cu113
pygcn==1.1.3
pyparsing==3.0.8
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
rdflib==6.2.0
requests==2.28.2
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.10.0
seaborn==0.12.2
sentry-sdk==1.5.10
setproctitle==1.2.3
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
statsmodels==0.13.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.6.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.4.8
timm==0.6.7
torch==1.12.1+cu113
torch-geometric==2.2.0
torch-scatter==2.1.0+pt112cu113
torch-sparse==0.6.16+pt112cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.64.0
typing-extensions==3.10.0.0
urllib3==1.26.14
uWSGI @ file:///croot/uwsgi_1679988297904/work
wandb==0.12.14
Werkzeug==2.3.6
wrapt==1.12.1
xgboost==1.6.0
zipp==3.8.0
